+++ src/extension.ts
@@ -1,9 +1,11 @@
 "use strict";
 
 import * as vscode from "vscode";
+import { PhpGenerator } from "../out/generators/php";
 import { Provider } from "./provider";
 import {
     IBaseProvider,
+    JavaProvider,
     JsonProvider,
     PhpProvider,
     PythonProvider,
@@ -14,12 +16,12 @@ import {
 function goToDefinition(range: vscode.Range) {
     const editor: vscode.TextEditor = vscode.window.activeTextEditor;
 
-        // Center the method in the document
-        editor.revealRange(range, vscode.TextEditorRevealType.InCenter);
-        // Select the method name
-        editor.selection = new vscode.Selection(range.start, range.end);
-        // Swap the focus to the editor
-        vscode.window.showTextDocument(editor.document, editor.viewColumn, false);
+    // Center the method in the document
+    editor.revealRange(range, vscode.TextEditorRevealType.InCenter);
+    // Select the method name
+    editor.selection = new vscode.Selection(range.start, range.end);
+    // Swap the focus to the editor
+    vscode.window.showTextDocument(editor.document, editor.viewColumn, false);
 }
 
 export function activate(context: vscode.ExtensionContext) {
@@ -41,6 +43,10 @@ export function activate(context: vscode.ExtensionContext) {
         providers.push(new JsonProvider());
     }
 
+    if (allowedProviders.length === 0 || allowedProviders.indexOf("java") !== -1) {
+        providers.push(new JavaProvider());
+    }
+
     if (allowedProviders.length === 0 || allowedProviders.indexOf("openhab") !== -1) {
         providers.push(new RuleProvider());
     }
@@ -51,8 +57,31 @@ export function activate(context: vscode.ExtensionContext) {
 
     const provider = new Provider(providers as Array<IBaseProvider<string | vscode.TreeItem>>);
 
+    if (vscode.window.activeTextEditor.document !== null) {
+        provider.refresh(vscode.window.activeTextEditor.document);
+    }
+
     vscode.window.registerTreeDataProvider("tree-outline", provider);
     vscode.commands.registerCommand("extension.treeview.goto", (range: vscode.Range) => goToDefinition(range));
+    vscode.commands.registerCommand("extension.treeview.extractInterface", (a: vscode.TreeItem) => {
+        provider.getTokenTree().then((tokenTree) => {
+            tokenTree.classes.map((t) => {
+                if (t.name === a.label.replace("@", "")) {
+                    provider.generateEntity(t, false, tokenTree.namespace, tokenTree.strict);
+                }
+            });
+        });
+    });
+
+    vscode.commands.registerCommand("extension.treeview.implementInterface", (a: vscode.TreeItem) => {
+        provider.getTokenTree().then((tokenTree) => {
+            tokenTree.interfaces.map((t) => {
+                if (t.name === a.label) {
+                    provider.generateEntity(t, true, tokenTree.namespace, tokenTree.strict);
+                }
+            });
+        });
+    });
 }
 
 export function deactivate() {