+++ src/providers/typescript.ts
@@ -1,3 +1,5 @@
+import * as fs from "fs";
+import * as os from "os";
 import * as ts from "typescript-parser";
 import * as vscode from "vscode";
 import { Provider } from "./../provider";
@@ -7,7 +9,7 @@ import { IBaseProvider } from "./base";
 export class TypescriptProvider implements IBaseProvider<vscode.TreeItem> {
     private config: vscode.WorkspaceConfiguration;
     private parser: ts.TypescriptParser;
-    private tree: token.ITokenTree = {} as token.ITokenTree;
+    private tree: Thenable<token.ITokenTree>;
 
     private readonly VISIBILITY = [
         "private", "protected", "public",
@@ -18,82 +20,28 @@ export class TypescriptProvider implements IBaseProvider<vscode.TreeItem> {
         this.config = vscode.workspace.getConfiguration("treeview.js");
     }
 
-    get roChar(): string {
-        return this.config.has("readonlyCharacter") ?
-        this.config.get("readonlyCharacter") : "@";
-    }
-
     public hasSupport(langId: string): boolean {
         return langId.toLowerCase() === "typescript" ||
             langId.toLowerCase() === "javascript";
     }
 
-    public refresh(event?: vscode.TextDocumentChangeEvent): void {
+    public refresh(document: vscode.TextDocument): void {
         this.config = vscode.workspace.getConfiguration("treeview.js");
-        // console.log("TypeScript/JavaScript Tree View provider refresh triggered")
-    }
+        const useStrict = document.getText().toString().substr(1, 10) === "use strict";
 
-    public getTokenTree(): Thenable<token.ITokenTree> {
-        const text = vscode.window.activeTextEditor.document.getText();
-        const useStrict = text.toString().substr(1, 10) === "use strict";
-
-        return this.parser.parseSource(text).then((raw: ts.File) => {
+        this.tree = this.parser.parseSource(document.getText()).then((raw: ts.File) => {
             const tree = {} as token.ITokenTree;
             tree.strict = useStrict;
 
+            for (const ns of raw.resources) {
+                if (ns instanceof ts.Namespace || ns instanceof ts.Module) {
+                    for (const dec of ns.declarations) {
+                        this.walk(dec, tree, ns.name);
+                    }
+                }
+            }
             for (const dec of raw.declarations) {
-                if (tree.nodes === undefined) {
-                    tree.nodes = [];
-                }
-
-                if (dec instanceof ts.ClassDeclaration || dec instanceof ts.InterfaceDeclaration) {
-                    if (dec instanceof ts.ClassDeclaration && dec.ctor !== undefined) {
-                        dec.ctor.name = "constructor";
-                        dec.methods.unshift(dec.ctor as ts.MethodDeclaration);
-                    }
-
-                    tree.nodes.push({
-                        methods: this.handleMethods(dec.methods),
-                        name: dec.name,
-                        properties: this.handleProperties(dec.properties),
-                        visibility: dec.isExported === true ? "public" : "protected",
-                    } as token.IEntityToken);
-                }
-
-                if (dec instanceof ts.VariableDeclaration) {
-                    const startPosition = vscode.window.activeTextEditor.document.positionAt(dec.start);
-
-                    if (tree.variables === undefined) {
-                        tree.variables = [];
-                    }
-
-                    tree.variables.push({
-                        name: `${dec.isConst ? this.roChar : ""}${dec.name}`,
-                        position: this.generateRangeForSelection(dec.name, dec.start),
-                        type: dec.type === undefined ? "any" : dec.type,
-                        visibility: dec.isExported === true ? "public" : "protected",
-                    } as token.IVariableToken);
-                }
-
-                if (dec instanceof ts.FunctionDeclaration) {
-                    const startPosition = vscode.window.activeTextEditor.document.positionAt(dec.start);
-
-                    if (tree.functions === undefined) {
-                        tree.functions = [];
-                    }
-
-                    tree.functions.push({
-                        arguments: this.handleArguments(dec.parameters),
-                        name: dec.name,
-                        position: new vscode.Range(
-                            startPosition,
-                            new vscode.Position(startPosition.line, startPosition.character),
-                        ),
-                        static: true,
-                        type: dec.type === null ? "any" : dec.type,
-                        visibility: dec.isExported === true ? "public" : "protected",
-                    } as token.IMethodToken);
-                }
+                this.walk(dec, tree);
             }
 
             for (const imp of raw.imports) {
@@ -132,12 +80,267 @@ export class TypescriptProvider implements IBaseProvider<vscode.TreeItem> {
         });
     }
 
+    public getTokenTree(): Thenable<token.ITokenTree> {
+        return this.tree;
+    }
+
     public getTreeItem(element: vscode.TreeItem): vscode.TreeItem | Thenable<vscode.TreeItem> {
         return element;
     }
 
     public getChildren(element?: vscode.TreeItem): Thenable<vscode.TreeItem[]> {
         return Promise.resolve([]);
+    }
+
+    public getDocumentName(entityName: string, includeBodies: boolean = false): Thenable<string> {
+        return vscode.window.showQuickPick([
+            new QuickPickItem("JavaScript", "Will create a `.js` file", "js"),
+            new QuickPickItem("TypeScript", "Will create a `.ts` file", "ts"),
+        ], {
+            ignoreFocusOut: true,
+            placeHolder: "Chose file extension",
+        }).then((r: QuickPickItem) => {
+            let name = entityName;
+            if (name.indexOf(".") !== -1) {
+                const nsSplit = name.split(".");
+                name = nsSplit.pop();
+            }
+
+            return (includeBodies ?
+                `${name}.${r.detail}` : `I${name}.ts`);
+        });
+    }
+
+    public generate(
+        entityName: string,
+        skeleton: (token.IInterfaceToken | token.IClassToken),
+        includeBodies: boolean,
+        options: any = {},
+    ): vscode.TextEdit[] {
+        if (entityName.indexOf(".") !== -1) {
+            const nsSplit = entityName.split(".");
+            entityName = nsSplit.pop();
+            options.ns = nsSplit.join(".");
+        }
+        const hasNs = (options.ns !== undefined && options.ext === "ts");
+
+        const edits: vscode.TextEdit[] = [];
+
+        if (options.strict !== undefined && options.strict === true) {
+            edits.push(new vscode.TextEdit(
+                new vscode.Range(
+                    new vscode.Position(edits.length, 0),
+                    new vscode.Position(edits.length, 13),
+                ),
+                "\"use strict\"" + os.EOL,
+            ));
+        }
+
+        if (hasNs) {
+            edits.push(new vscode.TextEdit(
+                new vscode.Range(
+                    new vscode.Position(edits.length, 0),
+                    new vscode.Position(edits.length, 1024),
+                ),
+                `export ${this.config.get("defaultNamespaceType")} ${options.ns} {` + os.EOL,
+            ));
+        }
+
+        edits.push(new vscode.TextEdit(
+            new vscode.Range(
+                new vscode.Position(edits.length, 0),
+                new vscode.Position(edits.length, 1024),
+            ),
+            (hasNs ? " ".repeat(4) : "") +
+            `export ${!includeBodies ? "interface" : `${skeleton.abstract ? "abstract " : ""}class`}` +
+                `${entityName} {` + os.EOL,
+        ));
+
+        if (skeleton.properties !== undefined) {
+            const properties = skeleton.properties.filter((c) => c.visibility === "public");
+            for (const constant of properties) {
+                const line = (hasNs ? " ".repeat(4) : "") + `    ` +
+                    `${includeBodies ? "public " : ""}${constant.name}` +
+                    `${constant.value !== "" ? `= ${constant.value}` : ""};`;
+
+                edits.push(new vscode.TextEdit(
+                    new vscode.Range(
+                        new vscode.Position(edits.length, 0),
+                        new vscode.Position(edits.length, line.length),
+                    ),
+                    line + os.EOL,
+                ));
+
+                if (properties.indexOf(constant) === properties.length - 1 &&
+                    skeleton.methods.length !== 0) {
+                    const constantPosition = skeleton.constants.indexOf(constant);
+                    edits.push(new vscode.TextEdit(
+                        new vscode.Range(
+                            new vscode.Position(edits.length, 0),
+                            new vscode.Position(edits.length, 1),
+                        ),
+                        os.EOL,
+                    ));
+                }
+            }
+        }
+
+        if (skeleton.methods !== undefined) {
+            const methods = skeleton.methods.filter((m) => m.visibility === "public");
+            for (const method of methods) {
+                if (!includeBodies && method.static) {
+                    continue;
+                }
+                let body = ";";
+                if (includeBodies) {
+                    body = (hasNs ? " ".repeat(4) : "") +
+                    `${os.EOL}    {` +
+                    (hasNs ? " ".repeat(4) : "") +
+                    `        throw new Error(\"Not implemented\");` +
+                    (hasNs ? " ".repeat(4) : "") +
+                    `${os.EOL}    }` + (methods.indexOf(method) === methods.length - 1 ? "" : os.EOL);
+                }
+
+                const args: string[] = [];
+                for (const arg of method.arguments) {
+                    args.push(
+                        `${arg.name}: ${arg.type}${arg.value !== "" ? ` = ${arg.value}` : ""}`,
+                    );
+                }
+                const returnType: string = method.type !== undefined && method.type !== "mixed" ?
+                    method.type : "";
+
+                const line = (hasNs ? " ".repeat(4) : "") +
+                    `    ${includeBodies ? `public ${skeleton.abstract ? "abstract " : ""}` +
+                        `${method.static ? "static " : ""}` : ""}` +
+                        `${method.name}(${args.join(", ")})` +
+                        `${returnType !== "" ? `: ${returnType}` : ""}${body}`;
+
+                edits.push(new vscode.TextEdit(
+                    new vscode.Range(
+                        new vscode.Position(edits.length + (includeBodies ? 2 : 0), 0),
+                        new vscode.Position(edits.length + (includeBodies ? 2 : 0), line.length),
+                    ),
+                    line + os.EOL,
+                ));
+            }
+        }
+
+        edits.push(new vscode.TextEdit(
+            new vscode.Range(
+                new vscode.Position(edits.length + (includeBodies ? 2 : 0), 0),
+                new vscode.Position(edits.length + (includeBodies ? 2 : 0), 1024),
+            ),
+            (hasNs ? " ".repeat(4) : "") + "}" + os.EOL,
+        ));
+
+        if (hasNs) {
+            edits.push(new vscode.TextEdit(
+                new vscode.Range(
+                    new vscode.Position(edits.length, 0),
+                    new vscode.Position(edits.length, 1024),
+                ),
+                "}",
+            ));
+        }
+
+        return edits;
+    }
+
+    private walk(dec: ts.Declaration, tree: token.ITokenTree, namespace?: string) {
+        if (dec instanceof ts.ClassDeclaration) {
+            if (tree.classes === undefined) {
+                tree.classes = [];
+            }
+
+            if (dec instanceof ts.ClassDeclaration && dec.ctor !== undefined) {
+                dec.ctor.name = "constructor";
+                dec.methods.unshift(dec.ctor as ts.MethodDeclaration);
+            }
+
+            let entityName = (namespace !== undefined ? `${namespace}.` : "") + `${dec.name}`;
+            if (this.config.has("namespacePosition")) {
+                if (this.config.get("namespacePosition") === "suffix") {
+                    entityName =
+                        `${dec.name}${namespace !== undefined ? `: ${namespace}` : ""}`;
+                }
+
+                if (this.config.get("namespacePosition") === "none") {
+                    entityName = `${dec.name}`;
+                }
+            }
+
+            const def = vscode.window.activeTextEditor.document.getText(new vscode.Range(
+                this.offsetToPosition(dec.start),
+                this.offsetToPosition(dec.end),
+            )).split(" ").slice(0, 5);
+
+            tree.classes.push({
+                abstract: (def.indexOf("abstract") > -1),
+                methods: this.handleMethods(dec.methods),
+                name: entityName,
+                properties: this.handleProperties(dec.properties),
+                visibility: dec.isExported === true ? "public" : "protected",
+            } as token.IClassToken);
+        }
+
+        if (dec instanceof ts.InterfaceDeclaration) {
+            if (tree.interfaces === undefined) {
+                tree.interfaces = [];
+            }
+
+            let entityName = (namespace !== undefined ? `${namespace}.` : "") + `${dec.name}`;
+            if (this.config.has("namespacePosition")) {
+                if (this.config.get("namespacePosition") === "suffix") {
+                    entityName =
+                        `${dec.name}${namespace !== undefined ? `: ${namespace}` : ""}`;
+                }
+
+                if (this.config.get("namespacePosition") === "none") {
+                    entityName = `${dec.name}`;
+                }
+            }
+
+            tree.interfaces.push({
+                methods: this.handleMethods(dec.methods),
+                name: entityName,
+                properties: this.handleProperties(dec.properties)
+                    .filter((p) => p.visibility === "public"),
+                visibility: dec.isExported === true ? "public" : "protected",
+            } as token.IInterfaceToken);
+        }
+
+        if (dec instanceof ts.VariableDeclaration) {
+            const startPosition = vscode.window.activeTextEditor.document.positionAt(dec.start);
+
+            if (tree.variables === undefined) {
+                tree.variables = [];
+            }
+
+            tree.variables.push({
+                name: `${dec.name}`,
+                position: this.generateRangeForSelection(dec.name, dec.start),
+                type: dec.type === undefined ? "any" : dec.type,
+                visibility: dec.isExported === true ? "public" : "protected",
+            } as token.IVariableToken);
+        }
+
+        if (dec instanceof ts.FunctionDeclaration) {
+            const startPosition = vscode.window.activeTextEditor.document.positionAt(dec.start);
+
+            if (tree.functions === undefined) {
+                tree.functions = [];
+            }
+
+            tree.functions.push({
+                arguments: this.handleArguments(dec.parameters),
+                name: dec.name,
+                position: this.generateRangeForSelection(dec.name, dec.start),
+                static: true,
+                type: dec.type === null ? "any" : dec.type,
+                visibility: dec.isExported === true ? "public" : "protected",
+            } as token.IMethodToken);
+        }
     }
 
     private handleProperties(children: any[]): token.IPropertyToken[] {
@@ -150,7 +353,7 @@ export class TypescriptProvider implements IBaseProvider<vscode.TreeItem> {
             )).split(" ").slice(0, 5);
 
             properties.push({
-                name: (def.indexOf("readonly") > -1 ? this.roChar : "") + property.name,
+                name: property.name,
                 position: this.generateRangeForSelection(property.name, property.start),
                 readonly: (def.indexOf("readonly") > -1),
                 static: (def.indexOf("static") > -1),
@@ -202,6 +405,7 @@ export class TypescriptProvider implements IBaseProvider<vscode.TreeItem> {
             )).split(" ").slice(0, 5);
 
             methods.push({
+                abstract: (def.indexOf("abstract") > -1),
                 arguments: this.handleArguments(method.parameters),
                 name: method.name,
                 position: this.generateRangeForSelection(method.name, method.start),
@@ -250,4 +454,8 @@ export class TypescriptProvider implements IBaseProvider<vscode.TreeItem> {
             new vscode.Position(startPosition.line, startIndex),
         );
     }
+}
+
+class QuickPickItem implements vscode.QuickPickItem {
+    constructor(public label, public description, public detail?) {}
 }