+++ src/tokens/items/interface.ts
@@ -1,0 +1,5 @@
+import { BaseItem } from "./base";
+
+export class InterfaceItem extends BaseItem {
+    public contextValue: string = "interface";
+}