+++ src/tokens/items/section.ts
@@ -1,0 +1,4 @@
+import { BaseItem } from "./base";
+
+export class SectionItem extends BaseItem {
+}