+++ src/tokens/items/variable.ts
@@ -1,0 +1,5 @@
+import { BaseItem } from "./base";
+
+export class VariableItem extends BaseItem {
+    public contextValue: string = "variable";
+}