+++ src/tokens/tree/class.ts
@@ -1,0 +1,5 @@
+import { ITraitToken } from "./trait";
+
+export interface IClassToken extends ITraitToken {
+    traits?: ITraitToken[];
+}