+++ src/tokens/tree/constant.ts
@@ -1,0 +1,3 @@
+import { IVariableToken } from "./variable";
+
+export interface IConstantToken extends IVariableToken { }