+++ src/tokens/tree/function.ts
@@ -1,0 +1,5 @@
+import { IVariableToken } from "../";
+
+export interface IFunctionToken extends IVariableToken {
+    arguments?: IVariableToken[];
+}