+++ src/tokens/tree/import.ts
@@ -1,0 +1,5 @@
+import { IBaseToken } from "./base";
+
+export interface ImportToken extends IBaseToken {
+    alias?: string;
+}