+++ src/tokens/tree/interface.ts
@@ -1,0 +1,13 @@
+import { IBaseToken } from "./base";
+import { IConstantToken } from "./constant";
+import { IMemberToken } from "./member";
+import { IMethodToken } from "./method";
+import { IPropertyToken } from "./property";
+
+export interface IInterfaceToken extends IBaseToken, IMemberToken {
+    constants?: IConstantToken[];
+    properties?: IPropertyToken[];
+    methods?: IMethodToken[];
+    readonly?: boolean;
+    abstract?: boolean;
+}