+++ src/tokens/tree/property.ts
@@ -1,0 +1,6 @@
+import { IMemberToken } from "./member";
+import { IVariableToken } from "./variable";
+
+export interface IPropertyToken extends IVariableToken, IMemberToken {
+    readonly?: boolean;
+}