+++ src/tokens/tree/type.ts
@@ -1,0 +1,3 @@
+import { IBaseToken } from "./base";
+
+export interface ITypeToken extends IBaseToken {}