+++ CHANGELOG.md
@@ -2,11 +2,32 @@
 
 _[BC]_ Stands for *B*reaking *C*hange
 
+## [2.0.0]
+
+### Notes
+
+- If you have configured a list of enabled providers, you must update your configuration to include `java` if you want to enable support for it
+
+### Added
+
+- [JAVA] - Initial Java support
+- [ALL] - Add context menu for interface creation and skeleton creation
+- [PHP/JS/TS] - Add handling for `abstract` and `final` keyword
+- [CONFIG] - Add `treeview.abstractCharacter` for consistent global configuration of `abstract` prefix
+- [JAVA] - Add `treeview.java.namespacePosition` for namespace positioning
+
+### Changes
+
+- [ALL] - Switch value objects for different tree items
+- [ALL] - Address issues with tree ghosting when all windows are closed or unsupported document is active
+- [ALL] - Centralized common code parts (icon, command, character prefixes)
+- [CONFIG] - Replaced `treeview.js.readonlyCharacter` with `treeview.readonlyCharacter` to consistently configure global symbols
+
 ## [1.2.0]
 
 ### Added
 
-- [PY] Initial python support
+- [PY] - Initial python support
 
 ## [1.1.0]